Add tox.ini
tox gives a simple way to handle creation of virtualenvs and
installation of necessary packages for running tests.

This allows you to call `tox` from TeamCity and not have to have a bunch
of extra build steps to prepare the virtualenvs.

As it stands in your book, the example doesn't work unless you have
Django installed in your system Python and I don't want to install it
there. I suggest adding a tox.ini and revising the book so that TeamCity
calls tox.